# evmrs: rewrite bench script in rust & add features