--- a/rust/bencher/Cargo.toml
+++ b/rust/bencher/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "bencher"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+chrono = "0.4.38"
+clap = { version = "4.5.20", features = ["derive"] }
+serde = { version = "1.0.213", features = ["derive"] }
+toml = "0.8.19"