- rewrite bench script in Rust and add features:
  - make it possible to have additive and subtractive features based on a baseline feature
    - providing a feature without prefix means it gets added to the baseline
    - providing a feature prefixed by `~` means it gets subtracted from the baseline
    - a feature map is saved to be able to later reproduce the benchmarks even if the features changed
  - add flags to customize runs, timeout and benchmark regex
- change tests script to run tests in non-full-mode but for `--no-default-features`, `--all-features` and `--features performance`

```
❯ ./scripts/bench.sh --help
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.09s
     Running `target/debug/bencher --help`
Run Go interpreter benchmarks for evmzero, lfvm and geth as well as for evmrs with different feature sets

Usage: bencher [OPTIONS] [FEATURES]...

Arguments:
  [FEATURES]...  evmrs features to benchmark. These features are relative to baseline. If they start with `~` they are subtracted and other wise added to baseline

Options:
      --evmrs-only             
      --runs <RUNS>            [default: 20]
      --timeout <TIMEOUT>      [default: 1h]
      --benchmark <BENCHMARK>  [default: ^Benchmark[a-zA-Z]+]
      --baseline <BASELINE>    [default: performance]
  -h, --help                   Print help
  -V, --version                Print version

```
```
❯ ./scripts/bench.sh --runs 1 ~mimalloc jumptable-tail-call
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 0.07s
     Running `target/debug/bencher --runs 1 '~mimalloc' jumptable-tail-call`
planned runs:
evmzero
lfvm
geth
evmrs with features: performance,~mimalloc = stack-array,custom-evmc,jumptable,hash-cache,code-analysis-cache
evmrs with features: performance,jumptable-tail-call = mimalloc,stack-array,custom-evmc,jumptable,hash-cache,code-analysis-cache,jumptable-tail-call
evmrs with features: performance = mimalloc,stack-array,custom-evmc,jumptable,hash-cache,code-analysis-cache
running make ...
running evmzero ...
running lfvm ...
running geth ...
running evmrs with features: performance,~mimalloc ...
running evmrs with features: performance,jumptable-tail-call ...
running evmrs with features: performance ...
goos: linux
goarch: amd64
pkg: github.com/Fantom-foundation/Tosca/go/integration_test/interpreter
cpu: Intel(R) Xeon(R) CPU E5-2690 v2 @ 3.00GHz
                   │    evmzero    │            evmrs#performance            │ evmrs#performance,jumptable-tail-call  │       evmrs#performance,~mimalloc       │                   geth                    │                  lfvm                  │
                   │    sec/op     │    sec/op      vs base                  │    sec/op      vs base                 │    sec/op      vs base                  │     sec/op       vs base                  │    sec/op      vs base                 │
Inc/1/                3.719µ ± ∞ ¹    6.689µ ± ∞ ¹         ~ (p=1.000 n=1) ²    6.218µ ± ∞ ¹        ~ (p=1.000 n=1) ²    6.986µ ± ∞ ¹         ~ (p=1.000 n=1) ²     16.928µ ± ∞ ¹         ~ (p=1.000 n=1) ²    5.073µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Inc/10/               3.692µ ± ∞ ¹    6.450µ ± ∞ ¹         ~ (p=1.000 n=1) ²    6.015µ ± ∞ ¹        ~ (p=1.000 n=1) ²    6.759µ ± ∞ ¹         ~ (p=1.000 n=1) ²     16.246µ ± ∞ ¹         ~ (p=1.000 n=1) ²    4.895µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Fib/1/                3.333µ ± ∞ ¹    5.788µ ± ∞ ¹         ~ (p=1.000 n=1) ²    5.346µ ± ∞ ¹        ~ (p=1.000 n=1) ²    6.090µ ± ∞ ¹         ~ (p=1.000 n=1) ²     15.749µ ± ∞ ¹         ~ (p=1.000 n=1) ²    4.261µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Fib/5/                9.172µ ± ∞ ¹   30.951µ ± ∞ ¹         ~ (p=1.000 n=1) ²   21.802µ ± ∞ ¹        ~ (p=1.000 n=1) ²   29.726µ ± ∞ ¹         ~ (p=1.000 n=1) ²     63.458µ ± ∞ ¹         ~ (p=1.000 n=1) ²   28.889µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Fib/10/               70.23µ ± ∞ ¹   327.34µ ± ∞ ¹         ~ (p=1.000 n=1) ²   222.87µ ± ∞ ¹        ~ (p=1.000 n=1) ²   314.13µ ± ∞ ¹         ~ (p=1.000 n=1) ²     610.44µ ± ∞ ¹         ~ (p=1.000 n=1) ²   311.47µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Fib/15/               712.6µ ± ∞ ¹   3313.1µ ± ∞ ¹         ~ (p=1.000 n=1) ²   2161.6µ ± ∞ ¹        ~ (p=1.000 n=1) ²   3186.8µ ± ∞ ¹         ~ (p=1.000 n=1) ²     6696.0µ ± ∞ ¹         ~ (p=1.000 n=1) ²   3431.8µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Fib/20/               7.521m ± ∞ ¹   36.411m ± ∞ ¹         ~ (p=1.000 n=1) ²   23.611m ± ∞ ¹        ~ (p=1.000 n=1) ²   35.116m ± ∞ ¹         ~ (p=1.000 n=1) ²     73.686m ± ∞ ¹         ~ (p=1.000 n=1) ²   38.068m ± ∞ ¹        ~ (p=1.000 n=1) ²
Sha3/1/               2.303µ ± ∞ ¹    3.211µ ± ∞ ¹         ~ (p=1.000 n=1) ²    3.197µ ± ∞ ¹        ~ (p=1.000 n=1) ²    3.139µ ± ∞ ¹         ~ (p=1.000 n=1) ²      8.833µ ± ∞ ¹         ~ (p=1.000 n=1) ²    1.339µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Sha3/10/              3.498µ ± ∞ ¹    5.558µ ± ∞ ¹         ~ (p=1.000 n=1) ²    5.646µ ± ∞ ¹        ~ (p=1.000 n=1) ²    5.463µ ± ∞ ¹         ~ (p=1.000 n=1) ²     21.973µ ± ∞ ¹         ~ (p=1.000 n=1) ²    4.140µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Sha3/100/             15.69µ ± ∞ ¹    30.06µ ± ∞ ¹         ~ (p=1.000 n=1) ²    30.07µ ± ∞ ¹        ~ (p=1.000 n=1) ²    28.40µ ± ∞ ¹         ~ (p=1.000 n=1) ²     142.62µ ± ∞ ¹         ~ (p=1.000 n=1) ²    29.89µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Sha3/1000/            156.6µ ± ∞ ¹    288.7µ ± ∞ ¹         ~ (p=1.000 n=1) ²    295.9µ ± ∞ ¹        ~ (p=1.000 n=1) ²    276.4µ ± ∞ ¹         ~ (p=1.000 n=1) ²     1342.8µ ± ∞ ¹         ~ (p=1.000 n=1) ²    301.7µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Arith/1/              4.309µ ± ∞ ¹    7.293µ ± ∞ ¹         ~ (p=1.000 n=1) ²    6.813µ ± ∞ ¹        ~ (p=1.000 n=1) ²    7.801µ ± ∞ ¹         ~ (p=1.000 n=1) ²     17.246µ ± ∞ ¹         ~ (p=1.000 n=1) ²    5.548µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Arith/10/             7.443µ ± ∞ ¹   17.245µ ± ∞ ¹         ~ (p=1.000 n=1) ²   14.670µ ± ∞ ¹        ~ (p=1.000 n=1) ²   17.681µ ± ∞ ¹         ~ (p=1.000 n=1) ²     35.377µ ± ∞ ¹         ~ (p=1.000 n=1) ²   15.929µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Arith/100/            36.95µ ± ∞ ¹   129.42µ ± ∞ ¹         ~ (p=1.000 n=1) ²   100.16µ ± ∞ ¹        ~ (p=1.000 n=1) ²   130.71µ ± ∞ ¹         ~ (p=1.000 n=1) ²     218.91µ ± ∞ ¹         ~ (p=1.000 n=1) ²   116.80µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Arith/280/            106.1µ ± ∞ ¹    335.4µ ± ∞ ¹         ~ (p=1.000 n=1) ²    271.3µ ± ∞ ¹        ~ (p=1.000 n=1) ²    334.1µ ± ∞ ¹         ~ (p=1.000 n=1) ²      589.5µ ± ∞ ¹         ~ (p=1.000 n=1) ²    324.6µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Memory/1/             5.592µ ± ∞ ¹   10.107µ ± ∞ ¹         ~ (p=1.000 n=1) ²    9.296µ ± ∞ ¹        ~ (p=1.000 n=1) ²   10.319µ ± ∞ ¹         ~ (p=1.000 n=1) ²     23.533µ ± ∞ ¹         ~ (p=1.000 n=1) ²    9.060µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Memory/10/            9.722µ ± ∞ ¹   22.595µ ± ∞ ¹         ~ (p=1.000 n=1) ²   18.773µ ± ∞ ¹        ~ (p=1.000 n=1) ²   22.116µ ± ∞ ¹         ~ (p=1.000 n=1) ²     46.763µ ± ∞ ¹         ~ (p=1.000 n=1) ²   23.115µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Memory/100/           50.42µ ± ∞ ¹   163.43µ ± ∞ ¹         ~ (p=1.000 n=1) ²   123.77µ ± ∞ ¹        ~ (p=1.000 n=1) ²   149.95µ ± ∞ ¹         ~ (p=1.000 n=1) ²     279.89µ ± ∞ ¹         ~ (p=1.000 n=1) ²   163.64µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Memory/1000/          444.0µ ± ∞ ¹   1344.8µ ± ∞ ¹         ~ (p=1.000 n=1) ²   1015.2µ ± ∞ ¹        ~ (p=1.000 n=1) ²   1280.7µ ± ∞ ¹         ~ (p=1.000 n=1) ²     2586.3µ ± ∞ ¹         ~ (p=1.000 n=1) ²   1560.8µ ± ∞ ¹        ~ (p=1.000 n=1) ²
Memory/10000/         4.130m ± ∞ ¹   13.093m ± ∞ ¹         ~ (p=1.000 n=1) ²    9.876m ± ∞ ¹        ~ (p=1.000 n=1) ²   12.435m ± ∞ ¹         ~ (p=1.000 n=1) ²     25.585m ± ∞ ¹         ~ (p=1.000 n=1) ²   15.352m ± ∞ ¹        ~ (p=1.000 n=1) ²
Analysis/jumpdest/   1897.0n ± ∞ ¹   2464.0n ± ∞ ¹         ~ (p=1.000 n=1) ²   2595.0n ± ∞ ¹        ~ (p=1.000 n=1) ²   2600.0n ± ∞ ¹         ~ (p=1.000 n=1) ²   119332.0n ± ∞ ¹         ~ (p=1.000 n=1) ²    825.7n ± ∞ ¹        ~ (p=1.000 n=1) ²
Analysis/stop/       1860.0n ± ∞ ¹   2486.0n ± ∞ ¹         ~ (p=1.000 n=1) ²   2515.0n ± ∞ ¹        ~ (p=1.000 n=1) ²   2604.0n ± ∞ ¹         ~ (p=1.000 n=1) ²   118641.0n ± ∞ ¹         ~ (p=1.000 n=1) ²    816.0n ± ∞ ¹        ~ (p=1.000 n=1) ²
Analysis/push1/      1854.0n ± ∞ ¹   2442.0n ± ∞ ¹         ~ (p=1.000 n=1) ²   2529.0n ± ∞ ¹        ~ (p=1.000 n=1) ²   2569.0n ± ∞ ¹         ~ (p=1.000 n=1) ²   132083.0n ± ∞ ¹         ~ (p=1.000 n=1) ²    817.6n ± ∞ ¹        ~ (p=1.000 n=1) ²
Analysis/push32/     1838.0n ± ∞ ¹   2460.0n ± ∞ ¹         ~ (p=1.000 n=1) ²   2530.0n ± ∞ ¹        ~ (p=1.000 n=1) ²   2582.0n ± ∞ ¹         ~ (p=1.000 n=1) ²   108771.0n ± ∞ ¹         ~ (p=1.000 n=1) ²    820.8n ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean               20.29µ          45.51µ        +124.29%                    39.40µ        +94.15%                    45.43µ        +123.86%                      177.9µ        +776.49%                    34.44µ        +69.70%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```
```
❯ cat feature-map 
performance,~mimalloc = stack-array,custom-evmc,jumptable,hash-cache,code-analysis-cache
performance,jumptable-tail-call = mimalloc,stack-array,custom-evmc,jumptable,hash-cache,code-analysis-cache,jumptable-tail-call
performance = mimalloc,stack-array,custom-evmc,jumptable,hash-cache,code-analysis-cache
```